move add-init-script
